requirements.txt
fastapi
uvicorn
requests
jinja2
python-multipart
itsdangerous
aiofiles
pycryptodome==3.20.0
brotli-asgi
pillow
